Split OpaqueType tests into their own spec
